+++ e2e/flows/custom-module-resolutions.e2e.2.js
@@ -3,6 +3,7 @@ import path from 'path';
 import chai, { expect } from 'chai';
 import Helper from '../e2e-helper';
 import * as fixtures from '../fixtures/fixtures';
+import NpmCiRegistry, { supportNpmCiRegistryTesting } from '../npm-ci-registry';
 
 chai.use(require('chai-fs'));
 
@@ -249,6 +250,92 @@ describe('custom module resolutions', function () {
               .to.have.property('utils/is-type')
               .that.equal(`${packageName}/utils/is-type.js`);
           });
+        });
+      });
+    });
+  });
+  describe('using custom module directory when a component uses an internal file of another component', () => {
+    const npmCiRegistry = new NpmCiRegistry(helper);
+    before(() => {
+      helper.setNewLocalAndRemoteScopes();
+      const bitJson = helper.readBitJson();
+      bitJson.resolveModules = { modulesDirectories: ['src'] };
+      helper.writeBitJson(bitJson);
+      npmCiRegistry.setCiScopeInBitJson();
+      helper.createFile('src/utils', 'is-type.js', '');
+      helper.createFile('src/utils', 'is-type-internal.js', fixtures.isType);
+      helper.addComponent('src/utils/is-type.js src/utils/is-type-internal.js', {
+        i: 'utils/is-type',
+        m: 'src/utils/is-type.js'
+      });
+
+      const isStringFixture =
+        "const isType = require('utils/is-type-internal');\n module.exports = function isString() { return isType() +  ' and got is-string'; };";
+      helper.createFile('src/utils', 'is-string.js', '');
+      helper.createFile('src/utils', 'is-string-internal.js', isStringFixture);
+      helper.addComponent('src/utils/is-string.js src/utils/is-string-internal.js', {
+        i: 'utils/is-string',
+        m: 'src/utils/is-string.js'
+      });
+
+      const barFooFixture =
+        "const isString = require('utils/is-string-internal');\n module.exports = function foo() { return isString() + ' and got foo'; };";
+      helper.createFile('src/bar', 'foo.js', barFooFixture);
+      helper.addComponent('src/bar/foo.js', { i: 'bar/foo', m: 'src/bar/foo.js' });
+    });
+    it('bit status should not warn about missing packages', () => {
+      const output = helper.runCmd('bit status');
+      expect(output).to.not.have.string('missing');
+    });
+    it('bit show should show the dependencies correctly', () => {
+      const output = helper.showComponentParsed('bar/foo');
+      expect(output.dependencies).to.have.lengthOf(1);
+      const dependency = output.dependencies[0];
+      expect(dependency.id).to.equal('utils/is-string');
+      expect(dependency.relativePaths[0].sourceRelativePath).to.equal('src/utils/is-string-internal.js');
+      expect(dependency.relativePaths[0].destinationRelativePath).to.equal('src/utils/is-string-internal.js');
+      expect(dependency.relativePaths[0].importSource).to.equal('utils/is-string-internal');
+      expect(dependency.relativePaths[0].isCustomResolveUsed).to.be.true;
+    });
+    describe('importing the component', () => {
+      before(() => {
+        helper.tagAllWithoutMessage();
+        helper.exportAllComponents();
+
+        helper.reInitLocalScope();
+        helper.addRemoteScope();
+        helper.importComponent('bar/foo');
+        fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), fixtures.appPrintBarFoo);
+      });
+      it('should generate the non-relative links correctly', () => {
+        const result = helper.runCmd('node app.js');
+        expect(result.trim()).to.equal('got is-type and got is-string and got foo');
+      });
+      it('should not show the component as modified', () => {
+        const output = helper.runCmd('bit status');
+        expect(output).to.not.have.string('modified');
+      });
+      (supportNpmCiRegistryTesting ? describe : describe.skip)('when dependencies are saved as packages', () => {
+        before(async () => {
+          await npmCiRegistry.init();
+          helper.importNpmPackExtension();
+          helper.removeRemoteScope();
+          npmCiRegistry.publishComponent('utils/is-type');
+          npmCiRegistry.publishComponent('utils/is-string');
+          npmCiRegistry.publishComponent('bar/foo');
+
+          helper.reInitLocalScope();
+          helper.runCmd('npm init -y');
+          helper.runCmd(`npm install @ci/${helper.remoteScope}.bar.foo`);
+        });
+        after(() => {
+          npmCiRegistry.destroy();
+        });
+        it('should be able to require its direct dependency and print results from all dependencies', () => {
+          const appJsFixture = `const barFoo = require('@ci/${helper.remoteScope}.bar.foo'); console.log(barFoo());`;
+          fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
+          const result = helper.runCmd('node app.js');
+          expect(result.trim()).to.equal('got is-type and got is-string and got foo');
         });
       });
     });