+++ src/environment/environment.js
@@ -58,7 +58,13 @@ export default class Environment {
     if (typeof bitId === 'string') {
       bitId = await BitId.parse(bitId, true);
     }
-    const componentsWithDependencies = await this.consumer.importComponents(BitIds.fromArray([bitId]));
+    const saveDependenciesAsComponents =
+      opts.saveDependenciesAsComponents === undefined ? true : opts.saveDependenciesAsComponents;
+    const componentsWithDependencies = await this.consumer.importComponents(
+      BitIds.fromArray([bitId]),
+      false,
+      saveDependenciesAsComponents
+    );
     const componentWithDependencies = componentsWithDependencies[0];
     const writeToPath = opts.writeToPath || this.path;
     const concreteOpts = {
@@ -70,7 +76,6 @@ export default class Environment {
       writeConfig: opts.conf,
       writeBitDependencies: opts.writeBitDependencies,
       createNpmLinkFiles: opts.createNpmLinkFiles,
-      saveDependenciesAsComponents: opts.saveDependenciesAsComponents !== false,
       writeDists: opts.dist,
       installNpmPackages: !!opts.installPackages, // convert to boolean
       installPeerDependencies: !!opts.installPackages, // convert to boolean